DB_test중

+++ Back-end/src/main/java/project/woori_saza/model/repo/QnaRepo.java
@@ -3,6 +3,6 @@ package project.woori_saza.model.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.woori_saza.model.domain.Qna;
 
-public interface QnaRepo extends JpaRepository<Qna,Long> {
+public interface QnaRepo extends JpaRepository<Qna,Integer> {
 
 }
