+++ Back-end/src/main/java/project/woori_saza/model/repo/ReviewRepo.java
@@ -3,6 +3,6 @@ package project.woori_saza.model.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.woori_saza.model.domain.Review;
 
-public interface ReviewRepo extends JpaRepository<Review,Integer> {
+public interface ReviewRepo extends JpaRepository<Review,Long> {
 
 }
